+++ src/host/configuration.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "ccf/ds/unit_strings.h"
+#include "ccf/pal/platform.h"
 #include "common/configuration.h"
 
 #include <optional>
@@ -15,23 +16,13 @@ namespace host
   {
     RELEASE,
     DEBUG,
-    VIRTUAL // Deprecated (use EnclavePlatform instead)
+    VIRTUAL // Deprecated (use ccf::pal::Platform instead)
   };
   DECLARE_JSON_ENUM(
     EnclaveType,
     {{EnclaveType::RELEASE, "Release"},
      {EnclaveType::DEBUG, "Debug"},
      {EnclaveType::VIRTUAL, "Virtual"}});
-
-  enum class EnclavePlatform
-  {
-    SGX,
-    SNP,
-    VIRTUAL,
-  };
-  DECLARE_JSON_ENUM(
-    EnclavePlatform,
-    {{EnclavePlatform::SNP, "SNP"}, {EnclavePlatform::VIRTUAL, "Virtual"}});
 
   enum class LogFormat
   {
@@ -65,7 +56,7 @@ namespace host
     {
       std::string file;
       EnclaveType type;
-      EnclavePlatform platform;
+      ccf::pal::Platform platform;
 
       bool operator==(const Enclave&) const = default;
     };
@@ -99,7 +90,6 @@ namespace host
 
     struct Logging
     {
-      ccf::LoggerLevel host_level = ccf::LoggerLevel::INFO;
       LogFormat format = LogFormat::TEXT;
 
       bool operator==(const Logging&) const = default;
@@ -174,7 +164,7 @@ namespace host
 
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::Logging);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::Logging);
-  DECLARE_JSON_OPTIONAL_FIELDS(CCHostConfig::Logging, host_level, format);
+  DECLARE_JSON_OPTIONAL_FIELDS(CCHostConfig::Logging, format);
 
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::Memory);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::Memory);