+++ tests/infra/utils.py
@@ -3,10 +3,9 @@
 import infra.path
 from hashlib import sha256
 import infra.snp as snp
-import infra.proc
 
 
-def get_measurement(enclave_type, enclave_platform, package, library_dir="."):
+def get_measurement(enclave_platform, package, library_dir="."):
     if enclave_platform == "virtual":
         return "Insecure hard-coded virtual measurement v1"
 
@@ -15,16 +14,14 @@ def get_measurement(enclave_type, enclave_platform, package, library_dir="."):
 
 
 def get_host_data_and_security_policy(
-    enclave_type, enclave_platform, package, library_dir="."
+    enclave_platform, package, library_dir=".", version=None
 ):
-    lib_path = infra.path.build_lib_path(
-        package, enclave_type, enclave_platform, library_dir
-    )
     if enclave_platform == "snp":
         security_policy = snp.get_container_group_security_policy()
         host_data = sha256(security_policy.encode()).hexdigest()
         return host_data, security_policy
     elif enclave_platform == "virtual":
+        lib_path = infra.path.build_lib_path(package, library_dir, version=version)
         hash = sha256(open(lib_path, "rb").read())
         return hash.hexdigest(), None
     else: